Add wrapper for sockrus initialization
First of all, sockrus hook has been moved into file `hook.go`. Second, commit
adds a wrapper for sockrus initialization, resp. to initialize logrus with
sockrus hook. This wrapper returns new logrus instance and ready-to-use
logrus.Entry with common pre-configured fields.
